memory bug with cell_data, now using space method for allocation
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/branches/concrete_parallel@10603 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/Serializer.cxx
@@ -109,7 +109,7 @@ Serializer::Deserialize(Phreeqc &phreeqc_ref, Dictionary &dictionary, std::vecto
 {
 	int ii = 0;
 	int dd = 0;
-	while (ii < ints.size())
+	while (ii < (int) ints.size())
 	{
 		PACK_TYPE type = (PACK_TYPE) ints[ii++];
 		switch (type)
